Split GameController methods from DevNestController
